+++ packages/config/src/configs/index.ts
@@ -38,6 +38,7 @@ import { polkadotAssetHubConfig } from './polkadotAssetHub';
 import { robonomicsConfig } from './robonomics';
 import { shidenConfig } from './shiden';
 import { subsocialAlphanetConfig } from './subsocialAlphanet';
+import { turingConfig } from './turing';
 import { turingAlphanetConfig } from './turingAlphanet';
 import { uniqueAlphaConfig } from './uniqueAlpha';
 
@@ -80,6 +81,7 @@ export const chainsConfigList: ChainConfig[] = [
   kusamaAssetHubConfig,
   polkadotAssetHubConfig,
   subsocialAlphanetConfig,
+  turingConfig,
   turingAlphanetConfig,
   uniqueAlphaConfig,
 ];