+++ packages/config/src/configs/turing.ts
@@ -1,0 +1,42 @@
+import {
+  BalanceBuilder,
+  ExtrinsicBuilder,
+  FeeBuilder,
+} from '@moonbeam-network/xcm-builder';
+import { movr, tur } from '../assets';
+import { moonriver, turing } from '../chains';
+import { AssetConfig } from '../types/AssetConfig';
+import { ChainConfig } from '../types/ChainConfig';
+
+export const turingConfig = new ChainConfig({
+  assets: [
+    new AssetConfig({
+      asset: tur,
+      balance: BalanceBuilder().system().account(),
+      destination: moonriver,
+      destinationFee: {
+        amount: FeeBuilder().assetManager().assetTypeUnitsPerSecond(),
+        asset: tur,
+      },
+      extrinsic: ExtrinsicBuilder()
+        .xTokens()
+        .transferMultiAsset(turing.parachainId)
+        .X1(),
+    }),
+    new AssetConfig({
+      asset: movr,
+      balance: BalanceBuilder().tokens().accounts(),
+      destination: moonriver,
+      destinationFee: {
+        amount: 0.00001,
+        asset: movr,
+      },
+      extrinsic: ExtrinsicBuilder().xTokens().transfer(),
+      fee: {
+        asset: tur,
+        balance: BalanceBuilder().system().account(),
+      },
+    }),
+  ],
+  chain: turing,
+});